+++ src/importer/treetagger/tests.rs
@@ -35,6 +35,20 @@ fn simple_token() {
 }
 
 #[test]
+fn encoding_latin() {
+    let mut importer = TreeTaggerImporter::default();
+    importer.encoding = Some("Latin1".into());
+    let actual = import_as_graphml_string(
+        importer,
+        Path::new("tests/data/import/treetagger/latin1"),
+        Some(TT_DEFAULT_VIS_CONFIG),
+    )
+    .unwrap();
+
+    assert_snapshot!(actual);
+}
+
+#[test]
 fn single_sentence() {
     let actual = import_as_graphml_string(
         TreeTaggerImporter::default(),